greenhouse-gql/api/pyproject.toml
[tool.poetry]
name = "greenhouse-gql"
version = "0.1.0"
description = ""
authors = ["mcsimps2@gmail.com"]

[tool.poetry.dependencies]
python = "3.9.5"
psycopg2 = "2.9.1"
psycopg2-binary = "2.9.1"
requests = "2.26.0"
flask = "2.0.2"
flask-sqlalchemy = "2.5.1"
flask-migrate = "3.1.0"
flask-bcrypt = "0.7.1"
graphene-sqlalchemy = "2.3.0"
flask-graphql = "2.0.1"
flask-jwt-extended = "4.3.1"
flask-cors = "3.0.10"
python-dateutil = "2.8.2"
pytz = "2022.1"
werkzeug = "2.0.3"
gunicorn = "20.1.0"

[tool.poetry.dev-dependencies]
black = "22.6.0"
pytest = "7.1.1"
pylint = "2.14.5"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
